fix: pro section and improvements
