+++ apps/docs/components/marketing/nextui-pro-section/index.ts
@@ -1,0 +1,1 @@
+export {NextUIProSection} from "./nextui-pro-section";
