+++ apps/docs/config/github-info.json
@@ -1,6 +1,6 @@
 {
-  "stars": { "raw": 22226, "formatted": "22.2K" },
-  "forks": 1566,
-  "subscribers": 83,
-  "openIssues": 406
+  "stars": { "raw": 22380, "formatted": "22.4K" },
+  "forks": 1580,
+  "subscribers": 85,
+  "openIssues": 428
 }
