+++ apps/docs/content/docs/guide/introduction.mdx
@@ -9,7 +9,7 @@ Welcome to the NextUI documentation!
 
 <NextImage
   priority
-  src="https://nextui.org/nextui-banner.png"
+  src="https://heroui-assets.nyc3.cdn.digitaloceanspaces.com/images/nextui-banner.jpeg"
   quality={100}
   width={700}
   height={350}
