+++ apps/docs/package.json
@@ -39,6 +39,7 @@
     "@nextui-org/use-clipboard": "workspace:*",
     "@nextui-org/use-infinite-scroll": "workspace:*",
     "@nextui-org/use-is-mobile": "workspace:*",
+    "clsx": "^1.2.1",
     "@radix-ui/react-scroll-area": "^1.0.5",
     "@react-aria/focus": "3.19.0",
     "@react-aria/i18n": "3.12.4",
